fix: rename webapp/fs_scans -> webapp/disk_scans (plugin name shadowing)
The webapp package `webapp/fs_scans` has the same leaf name as the
installed `fs_scans` plugin. On the Flask dev server (`flask run`, app
`run`) the app directory `/code/src/webapp` lands on `sys.path`, so the
plugin loader's `import_module('fs_scans')` resolved to OUR package
(which has no `list_pg_schemas`) instead of the site-packages plugin —
the loader hit `AttributeError`, logged "backend unreachable", and
silently disabled fs-scans. (Package-qualified imports via `webapp.run`
masked it, so earlier validation passed.)

Fix mirrors the existing `jobs` <-> `job_history` split: never name a
webapp package the same as the plugin it wraps. Renamed the package to
`webapp/disk_scans` (git mv preserves history) and updated all
`webapp.fs_scans` import paths in run.py / config_inspect.py / the
package itself. The `app.extensions['fs_scans']` key and the `FS_SCANS`
plugin descriptor (package `fs_scans`) are unchanged.

Verified via the exact dev-server import path (`from run import
create_app` in /code/src/webapp): is_enabled=True, 13 collections, 13
engines.
